requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
pylint==2.4.4
black==19.10b0
model_bakery==1.1.0